Adds success background task response test
