+++ test/integration/ggrc/models/test_background_task.py
@@ -3,10 +3,6 @@
 
 """Tests for background_task model."""
 
-import mock
-
-from ggrc.app import app
-
 from integration.ggrc import api_helper
 from integration.ggrc import TestCase
 from integration.ggrc.generator import ObjectGenerator
@@ -47,23 +43,6 @@ class TestBackgroundTask(TestCase):
         content.json['background_task']
     self.assertEqual(set(bg_task_content.keys()),
                      {"id", "selfLink", "status", "type"})
-
-  def test_bg_error_task(self):
-    """Test error BackgroundTask for success response"""
-    from ggrc.models.all_models import BackgroundTask
-    with mock.patch(
-        "ggrc.views.app.make_response",
-        return_value=app.make_response(
-            ("error", 500, [("Content-Type", "text/html")])
-        )
-    ):
-      self.api.client.post("/admin/create_missing_revisions")
-      task = BackgroundTask.query.first()
-      self.assertEqual(task.status, "Success")
-      self.assertEqual(
-          task.task_scheduled_response().status_code,
-          200
-      )
 
 
 class TestPermissions(TestCase):
