+++ test/unit/ggrc/integrations/test_taskqueue.py
@@ -9,6 +9,10 @@ import mock
 
 from ggrc.cloud_api.task_queue import get_app_engine_tasks
 
+from integration.ggrc.models import factories
+from ggrc.app import app
+from ggrc import db
+
 
 class TestTaskQueueAPI(unittest.TestCase):
   """Test class for TaskQueue api."""
@@ -57,3 +61,31 @@ class TestTaskQueueAPI(unittest.TestCase):
     ):
       tasks = get_app_engine_tasks("ggrc")
     self.assertEqual(tasks, ["reindex1542710631_29764"])
+
+  def test_bg_error_task(self):
+    """Test error BackgroundTask for success response"""
+    from ggrc.views import create_missing_revisions
+
+    with mock.patch(
+      "ggrc.views.app.make_response",
+      return_value=app.make_response(
+        ("error", 500, [("Content-Type", "text/html")])
+      )
+    ):
+      with mock.patch("ggrc.utils.revisions.do_missing_revisions"):
+        bg_task = factories.BackgroundTaskFactory(name="test_task_queue")
+        response = create_missing_revisions(bg_task)
+        self.assertEqual(response.status_code, 200)
+        db.session.delete(bg_task)
+        db.session.commit()
+
+  def test_bg_success_task(self):
+    """Test success BackgroundTask for success response"""
+    from ggrc.views import create_missing_revisions
+
+    with mock.patch("ggrc.utils.revisions.do_missing_revisions"):
+      bg_task = factories.BackgroundTaskFactory(name="test_task_queue")
+      response = create_missing_revisions(bg_task)
+      self.assertEqual(response.status_code, 200)
+      db.session.delete(bg_task)
+      db.session.commit()
